Reorganize sagent types into separate namespace.
Sagent export branch: sagent/export/main
copybarista-source-rev-sha256=28527d50041f9e8a74d4b61d359d369b4448236112a5ac6e496effa4125c0ced
copybarista-replay-base-sha256=0e67048818f59ea2f07b164e428dbd04fedb482dcc60670102b559a46ebe64de
copybarista-pr-state-version=1
