+++ sagent/agent/agent_test.py
@@ -57,18 +57,18 @@ from sagent.agent.state import (
     approx_tokens,
     tool_state_context,
 )
-from sagent.catalog.cost import (
-    PriceCatalog,
-    PriceCatalogProduct,
-    TokenCost,
-    TokenPrice,
-)
 from sagent.compaction.summary import SummaryCompactor
 from sagent.lib import last_models, token_count
 from sagent.lib.custom_json import JSON, json_freeze
 from sagent.providers import Google
 from sagent.tools.read import Read
 from sagent.types.compactor import CompactRestorable
+from sagent.types.cost import (
+    PriceCatalog,
+    PriceCatalogProduct,
+    TokenCost,
+    TokenPrice,
+)
 from sagent.types.model import (
     ModelLimits,
     ModelSpec,
