+++ sagent/agent/cost_tracker_test.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import pytest
 
 from sagent.agent.cost_tracker import CostTracker
-from sagent.catalog.cost import TokenCost
+from sagent.types.cost import TokenCost
 from sagent.types.model import ModelResponse, TokenCount
 from sagent.types.runtime import AssistantMessage
