+++ sagent/agent/session_io.py
@@ -47,9 +47,9 @@ from sagent.agent.context import (
     resolve_context,
 )
 from sagent.agent.state import ReadCacheEntry, ToolState
-from sagent.catalog.cost import TokenCost
 from sagent.lib.custom_json import FloatCodec, IntCodec
 from sagent.sessions import restrict_path
+from sagent.types.cost import TokenCost
 from sagent.types.model import Model, ModelRecipe, TokenCount
 from sagent.types.providers import ProviderOptions
 from sagent.types.runtime import (
