+++ sagent/agent/session_io_test.py
@@ -49,12 +49,12 @@ from sagent.agent.session_io import (
     unpersisted_session_error,
 )
 from sagent.agent.state import ReadCacheEntry, ToolState
-from sagent.catalog.cost import (
+from sagent.sessions import new_session_dir
+from sagent.types.cost import (
     PriceCatalog,
     PriceCatalogProduct,
     TokenPrice,
 )
-from sagent.sessions import new_session_dir
 from sagent.types.model import (
     ModelLimits,
     ModelRequest,
