+++ sagent/catalog/anthropic.py
@@ -16,11 +16,11 @@ from collections.abc import Mapping
 from types import MappingProxyType
 from typing import Final
 
-from sagent.catalog.capability import (
+from sagent.types.capability import (
     ModelCapability,
     ModelLimits,
 )
-from sagent.catalog.cost import (
+from sagent.types.cost import (
     PriceCatalog,
     PriceCatalogProduct,
     TokenPrice,
