+++ sagent/catalog/dashscope.py
@@ -13,16 +13,16 @@ from __future__ import annotations
 from collections.abc import Mapping
 from types import MappingProxyType
 
-from sagent.catalog.capability import (
+from sagent.types.capability import (
     ModelCapability,
     ModelLimits,
 )
-from sagent.catalog.cost import (
+from sagent.types.cost import (
     PriceCatalog,
     PriceCatalogProduct,
     TokenPrice,
 )
-from sagent.catalog.thinking import ThinkingEffort
+from sagent.types.thinking import ThinkingEffort
 
 
 __all__ = ["MODELS"]
