+++ sagent/catalog/google.py
@@ -14,16 +14,16 @@ from collections.abc import Mapping
 from types import MappingProxyType
 from typing import Final
 
-from sagent.catalog.capability import (
+from sagent.types.capability import (
     ModelCapability,
     ModelLimits,
 )
-from sagent.catalog.cost import (
+from sagent.types.cost import (
     PriceCatalog,
     PriceCatalogProduct,
     TokenPrice,
 )
-from sagent.catalog.thinking import ThinkingEffort
+from sagent.types.thinking import ThinkingEffort
 
 
 __all__ = ["API", "CLI", "MODELS", "SUBSCRIPTION"]
