+++ sagent/catalog/openai.py
@@ -16,16 +16,16 @@ from dataclasses import replace
 from types import MappingProxyType
 from typing import Final
 
-from sagent.catalog.capability import (
+from sagent.types.capability import (
     ModelCapability,
     ModelLimits,
 )
-from sagent.catalog.cost import (
+from sagent.types.cost import (
     PriceCatalog,
     PriceCatalogProduct,
     TokenPrice,
 )
-from sagent.catalog.thinking import ThinkingEffort
+from sagent.types.thinking import ThinkingEffort
 
 
 __all__ = ["API", "CHAT_MODELS", "MODELS", "SUBSCRIPTION"]
