+++ sagent/providers/selfhosted/server_test.py
@@ -11,7 +11,6 @@ from typing import cast
 
 import json
 
-from sagent.catalog.cost import PriceCatalogProduct
 from sagent.lib.custom_json import MutableJSON
 from sagent.providers.selfhosted.server import (
     SelfHostedModel,
@@ -28,6 +27,7 @@ from sagent.providers.selfhosted.server import (
     _tool_preamble,
     _tool_schema,
 )
+from sagent.types.cost import PriceCatalogProduct
 from sagent.types.model import ModelRequest
 from sagent.types.runtime import (
     AssistantMessage,
