+++ sagent/tools/agent_self_test.py
@@ -13,9 +13,9 @@ import pytest
 from sagent import providers as providers_module
 from sagent.agent.agent import Agent
 from sagent.agent.state import current_agent_var, tool_state_var
-from sagent.catalog.cost import TokenCost
 from sagent.testing import MockModelCaps
 from sagent.tools.agent_self import AgentSelf
+from sagent.types.cost import TokenCost
 from sagent.types.model import (
     ModelCapability,
     ModelLimits,
