+++ sagent/types/thinking.py
